+++ webviz_subsurface/plugins/_well_analysis/_views/_well_overview_view/_settings/_chart_type.py
@@ -1,0 +1,30 @@
+from typing import List
+
+import webviz_core_components as wcc
+from dash.development.base_component import Component
+from webviz_config.utils import StrEnum
+from webviz_config.webviz_plugin_subclasses import SettingsGroupABC
+
+from ...._types import ChartType
+
+
+class WellOverviewChartType(SettingsGroupABC):
+    class Ids(StrEnum):
+        CHARTTYPE = "charttype"
+
+    def __init__(self) -> None:
+        super().__init__("Chart Type")
+
+    def layout(self) -> List[Component]:
+        return [
+            wcc.RadioItems(
+                id=self.register_component_unique_id(self.Ids.CHARTTYPE),
+                options=[
+                    {"label": "Bar chart", "value": ChartType.BAR},
+                    {"label": "Pie chart", "value": ChartType.PIE},
+                    {"label": "Stacked area chart", "value": ChartType.AREA},
+                ],
+                value=ChartType.BAR,
+                vertical=True,
+            )
+        ]